now whenever a price is checked the price history is stored in a json file.  Created a script to generate a chart using all history.  Remove chart test script
